Refactoring part 2 (#36)
* Refactoring continued

* renamed test files to _whitebox_ where applicable

* Fixed naming

* Changes as discussed

* added nolint
