+++ main.go
@@ -6,8 +6,7 @@ SPDX-License-Identifier: Apache-2.0
 package main
 
 import (
-	"github.com/gardener/gardenctl-v2/pkg/cmd/root"
-
+	"github.com/gardener/gardenctl-v2/pkg/cmd"
 	gardencorev1beta1 "github.com/gardener/gardener/pkg/apis/core/v1beta1"
 	operationsv1alpha1 "github.com/gardener/gardener/pkg/apis/operations/v1alpha1"
 	utilruntime "k8s.io/apimachinery/pkg/util/runtime"
@@ -18,5 +17,5 @@ func main() {
 	utilruntime.Must(gardencorev1beta1.AddToScheme(scheme.Scheme))
 	utilruntime.Must(operationsv1alpha1.AddToScheme(scheme.Scheme))
 
-	root.Execute()
+	cmd.Execute()
 }
