+++ pkg/cmd/cmd_suite_test.go
@@ -4,7 +4,7 @@ SPDX-FileCopyrightText: 2021 SAP SE or an SAP affiliate company and Gardener con
 SPDX-License-Identifier: Apache-2.0
 */
 
-package root_test
+package cmd_test
 
 import (
 	"testing"
