+++ pkg/cmd/version/version.go
@@ -9,6 +9,7 @@ package version
 import (
 	"encoding/json"
 	"fmt"
+	"github.com/gardener/gardenctl-v2/pkg/cmd/base"
 
 	"github.com/gardener/gardenctl-v2/internal/util"
 
@@ -17,8 +18,8 @@ import (
 	"k8s.io/component-base/version"
 )
 
-// NewCommand returns a new version command.
-func NewCommand(f util.Factory, o *Options) *cobra.Command {
+// NewCmdVersion returns a new version command.
+func NewCmdVersion(f util.Factory, o *VersionOptions) *cobra.Command {
 	cmd := &cobra.Command{
 		Use:   "version",
 		Short: "Print the gardenctl version information",
@@ -30,7 +31,7 @@ func NewCommand(f util.Factory, o *Options) *cobra.Command {
 				return err
 			}
 
-			return runCommand(o)
+			return runCmdVersion(o)
 		},
 	}
 
@@ -40,7 +41,7 @@ func NewCommand(f util.Factory, o *Options) *cobra.Command {
 	return cmd
 }
 
-func runCommand(opt *Options) error {
+func runCmdVersion(opt *VersionOptions) error {
 	versionInfo := version.Get()
 
 	switch opt.Output {
@@ -71,6 +72,40 @@ func runCommand(opt *Options) error {
 		// There is a bug in the program if we hit this case.
 		// However, we follow a policy of never panicking.
 		return fmt.Errorf("options were not validated: --output=%q should have been rejected", opt.Output)
+	}
+
+	return nil
+}
+
+// VersionOptions is a struct to support version command
+// nolint
+type VersionOptions struct {
+	base.Options
+
+	// Short indicates if just the version number should be printed
+	Short bool
+	// Output defines the output format of the version information. Either 'yaml' or 'json'
+	Output string
+}
+
+// NewVersionOptions returns initialized VersionOptions
+func NewVersionOptions(ioStreams util.IOStreams) *VersionOptions {
+	return &VersionOptions{
+		Options: base.Options{
+			IOStreams: ioStreams,
+		},
+	}
+}
+
+// Complete adapts from the command line args to the data required.
+func (o *VersionOptions) Complete(f util.Factory, cmd *cobra.Command, args []string) error {
+	return nil
+}
+
+// Validate validates the provided options
+func (o *VersionOptions) Validate() error {
+	if o.Output != "" && o.Output != "yaml" && o.Output != "json" {
+		return fmt.Errorf(`--output must be either 'yaml' or 'json'`)
 	}
 
 	return nil
