+++ pkg/cmd/version/version_test.go
@@ -14,10 +14,21 @@ import (
 	. "github.com/onsi/gomega"
 )
 
-var _ = Describe("Options", func() {
+var _ = Describe("Command", func() {
+	It("should print version", func() {
+		streams, _, out, _ := util.NewTestIOStreams()
+		o := NewVersionOptions(streams)
+		cmd := NewCmdVersion(&util.FactoryImpl{}, o)
+
+		Expect(cmd.RunE(cmd, nil)).To(Succeed())
+		Expect(out.String()).To(ContainSubstring("GitVersion"))
+	})
+})
+
+var _ = Describe("VersionOptions", func() {
 	It("should validate", func() {
 		streams, _, _, _ := util.NewTestIOStreams()
-		o := NewOptions(streams)
+		o := NewVersionOptions(streams)
 		err := o.Validate()
 		Expect(err).ToNot(HaveOccurred())
 	})
